First POC
This is my first take at what might be the `isotope` API.

React and an array of options are passed. Each option is an object with
a test regexp and `key` and `value` functions which are passed the
prop name and value for each matching prop and expected to return a
string representing the key and value respectively.